Make odb handle more general (#266)
